添加typescript支持

+++ src/utils/AuthHelpers.ts
@@ -1,4 +1,8 @@
 import {useUserStore} from "@/store/userStore";
+import router from "@/router";
+import Configs from "@/config/Configs";
+import {ToastConfirm} from "@/utils/helpers";
+import {IUserInfo} from "@/api/modules/User";
 
 const tokenKey = "store_auth_token_key"
 const refreshTokenKey = "store_auth_refresh_token_key"
@@ -8,23 +12,23 @@ export default {
     getToken() {
         return localStorage.getItem(tokenKey)
     },
-    setToken(value) {
+    setToken(value: string) {
         localStorage.setItem(tokenKey, value)
     },
     getRefreshToken() {
         return localStorage.getItem(refreshTokenKey)
     },
-    setRefreshToken(value) {
+    setRefreshToken(value: string) {
         localStorage.setItem(refreshTokenKey, value)
     },
-    getUserinfo(){
+    getUserinfo(): IUserInfo | null {
         try {
             return JSON.parse(localStorage.getItem(userinfoKey) || "") || null;
         } catch (e) {
             return null
         }
     },
-    setUserinfo(value) {
+    setUserinfo(value: IUserInfo) {
         localStorage.setItem(userinfoKey, JSON.stringify(value));
     },
     syncUserinfo(isRefresh = false) { // 同步用户信息到store中
@@ -36,7 +40,16 @@ export default {
         token && userStore.setToken(token);
         refreshToken && userStore.setRefreshToken(refreshToken);
     },
-    removeUserInfo() {
+    async removeUserinfo(isReplace = true, showConfirm = false) {
+        if (showConfirm) {
+            const can = await ToastConfirm("确定退出登录吗?")
+            if (!can) {
+                return
+            }
+        }
         localStorage.clear();
+        if (isReplace) {
+            await router.replace({name: Configs.loginRouteName})
+        }
     }
 }
